update events filters
